% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{complete_matrix}
\alias{complete_matrix}
\title{Make a matrix symmetric by averaging with its transpose}
\usage{
complete_matrix(m, na.rm = TRUE)
}
\arguments{
\item{m}{Numeric matrix to complete.}

\item{na.rm}{Logical. Should missing values be dropped? (default: \code{TRUE})}
}
\value{
A completed matrix.
}
\description{
Makes a matrix symmetric by averaging the elements of the matrix and its
transpose. When This function fills in \code{NA} elements of a matrix with the corresponding
value from the matrix transpose, if available
}
\examples{
predictors <- c('auto', 'skill_var', 'task_var', 'task_sig', 'task_id',
'fb_job', 'job_comp', 'interdep', 'fb_others', 'soc_support')
m <- jobchar$sevar_r[c('perform', predictors), c('perform', predictors)]
complete_matrix(m)
}
