% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\encoding{UTF-8}
\name{configural_news}
\alias{configural_news}
\title{Retrieve the NEWS file for the configural package}
\usage{
configural_news()
}
\description{
This function gives a shortcut to the \code{utils::news(package = "configural")} function and displays configural's NEWS file, which contains version information, outlines additions and changes to the package, and describes other updates.
}
\examples{
configural_news()
}
