% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cor_covariance_meta}
\alias{cor_covariance_meta}
\title{Estimate the asymptotic sampling covariance matrix for the unique elements of
a meta-analytic correlation matrix}
\usage{
cor_covariance_meta(R, n, sevar, source = NULL, n_overlap = NULL)
}
\arguments{
\item{R}{A meta-analytic correlation matrix (can be full or lower-triangular).}

\item{n}{A matrix of total sample sizes for the meta-analytic correlations in \code{R} (can be full or lower-triangular).}

\item{sevar}{A matrix of estimated sampling error variances for the meta-analytic correlations in \code{R} (can be full or lower-triangular).}

\item{source}{A matrix indicating the sources of the meta-analytic correlations in \code{R} (can be full or lower-triangular). Used to estimate overlapping sample size for correlations when \code{n_overlap == NULL}.}

\item{n_overlap}{A matrix indicating the overlapping sample size for the unique (lower triangular) values in \code{R} (can be full or lower-triangular). Values must be arranged in the order returned by \code{cor_labels(colnames(R))}.}
}
\value{
The estimated asymptotic sampling covariance matrix
}
\description{
Estimate the asymptotic sampling covariance matrix for the unique elements of
a meta-analytic correlation matrix
}
\details{
If both \code{source} and \code{n_overlap} are \code{NULL}, it is assumed that all meta-analytic correlations come from the the same source.
}
\examples{
cor_covariance_meta(mindfulness$r, mindfulness$n, mindfulness$sevar_r, mindfulness$source)
}
\references{
Nel, D. G. (1985).
A matrix derivation of the asymptotic covariance matrix of sample correlation coefficients.
\emph{Linear Algebra and Its Applications, 67}, 137–145. \url{https://doi.org/10/c75jmg}

Wiernik, B. M. (2018).
\emph{Accounting for dependency in meta-analytic structural equations modeling: A flexible alternative to generalized least squares and two-stage structural equations modeling.}
Unpublished manuscript.
}
