% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\encoding{UTF-8}
\name{cor_labels}
\alias{cor_labels}
\title{Generate labels for correlations from a vector of variable names}
\usage{
cor_labels(var_names)
}
\arguments{
\item{var_names}{A character vector of variable names}
}
\value{
A vector of correlation labels
}
\description{
This function returns a vector of labels for the unique correlations between pairs of variables from a supplied vector of variable names
}
\examples{
cor_labels(colnames(mindfulness$r))
}
