% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{\%&\%}
\alias{\%&\%}
\title{Quadratic form matrix product}
\usage{
A \%&\% x
}
\arguments{
\item{A}{A square matrix}

\item{x}{A vector or matrix}
}
\value{
The quadratic product
}
\description{
Calculate the quadratic form \deqn{Q=x^{\prime}Ax}{Q = t(x) \%\*\% A \%\*\% x}
}
\examples{
diag(5) \%&\% 1:5
}
