% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\encoding{UTF-8}
\name{gre}
\alias{gre}
\title{Meta-analytic correlations of Graduate Record Examination subtests with graduate grade point average}
\format{list with entries \code{r} (mean oberseved correlations), \code{rho} (mean
corrected correlations), \code{n} (sample sizes), \code{sevar_r} (sampling error
variances for mean observed correlations), \code{sevar_rho} (sampling error
variances for mean corrected correlations), and \code{source} (character labels
indicating which meta-analytic correlations came from the same source)}
\usage{
data(jobchar)
}
\description{
Correlations between GRE subtests adn graduate student GPA from Kuncel et al. (2001).
}
\details{
GRE–GPA correlations in \code{rho} are corrected for direct range restriction on
the GRE and unreliability in GPA. Subtest intercorrelations in \code{rho} are
observed correlations computed among applicant norm samples. GRE subtest
intercorrelations in \code{r} are the values in \code{rho} attenuated using the bivariate
direct range restriction formula (Alexander et al., 1987). \code{sevar_rho} for
GRE subtest intercorrelations computed with an assumed
\if{latex}{\eqn{SD_\rho}}\ifelse{html}{\out{SD<sub>&rho;</sub>}}{SD_rho} = .02.
\code{sevar_r} for GRE subtest intercorrelations computed using the estimated
values for \code{r}, an assumed
\if{latex}{\eqn{SD_\rho}}\ifelse{html}{\out{SD<sub>&rho;</sub>}}{SD_rho} = .02,
and the N and k values for \code{rho}.
}
\examples{
data(gre)
}
\references{
Alexander, R. A., Carson, K. P., Alliger, G. M., & Carr, L. (1987).
Correcting doubly truncated correlations: An improved approximation for correcting the bivariate normal correlation when truncation has occurred on both variables.
\emph{Educational and Psychological Measurement, 47}(2), 309–315. \url{https://doi.org/10.1177/0013164487472002}

Kuncel, N. R., Hezlett, S. A., & Ones, D. S. (2001).
A comprehensive meta-analysis of the predictive validity of the graduate record examinations: Implications for graduate student selection and performance.
\emph{Psychological Bulletin, 127}(1), 162–181. \url{https://doi.org/10/cffkhc}
}
\keyword{datasets}
