% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{hrm}
\alias{hrm}
\title{Meta-analytic correlations of HRM practices with organizational financial performance}
\format{list with entries \code{r} (mean oberseved correlations), \code{rho} (mean
corrected correlations), \code{n} (sample sizes), \code{sevar_r} (sampling error
variances for mean observed correlations), \code{sevar_rho} (sampling error
variances for mean corrected correlations), and \code{source} (character labels
indicating which meta-analytic correlations came from the same source)}
\usage{
data(hrm)
}
\description{
Human resource management practice–organizational financial performance
correlations from Combs et al. (2006). Intercorrelations among HRM practices
from Guest et al. (2004).
}
\examples{
data(hrm)
}
\references{
Combs, J., Liu, Y., Hall, A., & Ketchen, D. (2006).
How much do high-performance work practices matter? A meta-analysis of their effects on organizational performance.
\emph{Personnel Psychology, 59}(3), 501–528. \url{https://doi.org/10/cm845n}

Guest, D., Conway, N., & Dewe, P. (2004).
Using sequential tree analysis to search for ‘bundles’ of HR practices.
\emph{Human Resource Management Journal, 14}(1), 79–96. \url{https://doi.org/10/ffxcfs}
}
\keyword{datasets}
