% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\encoding{UTF-8}
\name{jobchar}
\alias{jobchar}
\title{Meta-analytic correlations of job characteristics with performance and satisfaction}
\format{
list with entries \code{r} (mean observed correlations), \code{rho} (mean
corrected correlations), \code{n} (sample sizes), \code{sevar_r} (sampling error
variances for mean observed correlations), \code{sevar_rho} (sampling error
variances for mean corrected correlations), and \code{source} (character labels
indicating which meta-analytic correlations came from the same source)
}
\usage{
data(jobchar)
}
\description{
Self-rated job characteristics intercorrelations and correlations with
other-rated job performance and self-rated job satisfaction from Humphrey et
al. (2007).
}
\examples{
data(jobchar)
}
\references{
Humphrey, S. E., Nahrgang, J. D., & Morgeson, F. P. (2007).
Integrating motivational, social, and contextual work design features: A meta-analytic summary and theoretical extension of the work design literature.
\emph{Journal of Applied Psychology, 92}(5), 1332–1356. \doi{10.1037/0021-9010.92.5.1332}
}
\keyword{datasets}
