% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\encoding{UTF-8}
\name{prejudice}
\alias{prejudice}
\title{Correlations between study design moderators and effect sizes for prejudice reduction following intergroup contact}
\format{
list with entries \code{r} (observed correlations among moderators) and
\code{k} (number of samples in meta-analysis)
}
\usage{
data(prejudice)
}
\description{
Correlations among study design moderators and study design
moderator–observed prejudice reduction effect sizes from Pettigrew and Tropp
(2008). Note that correlations with effect size have been reverse-coded so
that a positive correlation indicates that a higher level of the moderator is
associated with \emph{larger} prejudice reduction.
}
\examples{
data(prejudice)
}
\references{
Pettigrew, T. F., & Tropp, L. R. (2006).
A meta-analytic test of intergroup contact theory.
\emph{Journal of Personality and Social Psychology, 90}(5), 751–783. \doi{10.1037/0022-3514.90.5.751}
}
\keyword{datasets}
