% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\encoding{UTF-8}
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Find the harmonic mean of a vector, matrix, or columns of a data.frame}
\usage{
harmonic_mean(x, na.rm = TRUE, zero = TRUE)
}
\arguments{
\item{x}{A vector, matrix, or data.frame}

\item{na.rm}{Logical. If \code{TRUE}, remove \code{NA} values before processing}

\item{zero}{Logical, If \code{TRUE}, if there are any zeroes, return 0, else, return the harmonic mean of the non-zero elements}
}
\value{
The harmonic mean of x
}
\description{
The harmonic mean is merely the reciprocal of the arithmetic mean of the reciprocals.
}
\examples{
harmonic_mean(1:10)
}
\author{
Adapted from \code{\link[psych:harmonic.mean]{psych::harmonic.mean()}} by William Revelle
}
