% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{transition}
\alias{transition}
\title{Calculate a transition matrix for a symmetric matrix}
\usage{
transition(p)
}
\arguments{
\item{p}{The number of columns in a matrix}
}
\description{
The transition matrix extracts the lower triangular elements from a vectorized symmetric matrix (Nel, 1985).
}
\examples{
transition(5)
}
\references{
Nel, D. G. (1985).
A matrix derivation of the asymptotic covariance matrix of sample correlation coefficients.
\emph{Linear Algebra and Its Applications, 67}, 137–145. \url{https://doi.org/10/c75jmg}
}
\author{
Based on internal functions from the \pkg{fungible} package by Niels Waller
}
\keyword{internal}
